requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Google ADK dependencies (adjust versions as needed)
google-adk>=0.1.0
google.genai>=0.1.0

# MongoDB dependencies
motor==3.3.2
pymongo==4.6.0
beanie==1.24.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AWS SDK for SES
boto3==1.34.0
botocore==1.34.0

# Email validation
email-validator==2.1.0

# Additional utilities
python-decouple==3.8
pydantic-settings==2.1.0

# Additional dependencies for men's health features
requests>=2.31.0
httpx>=0.25.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0